+++ src/measurement_package.h
@@ -12,7 +12,7 @@ public:
     RADAR
   } sensor_type_;
 
-  Eigen::VectorXd raw_measurements_;
+  Eigen::VectorXf raw_measurements_;
 };
 
 #endif /* MEASUREMENT_PACKAGE_H_ */
